organized menu routing folder structure
