+++ src/About/About.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import './About.css';
+
+class About extends Component {
+    constructor(props) {
+    //   super(props)
+    //   this.state = {
+    //     active: 'Home'
+    //   }
+    }
+
+    render() {
+        return (
+            <div>This is the About component</div>
+        )
+    }
+}
+
+export default About;
