+++ src/App/App.js
@@ -1,6 +1,5 @@
 import React, {Component} from 'react';
-import './App.css';
-import Menu from './Menu';
+import Menu from '../Menu/Menu';
 
 class App extends Component {
   constructor(props) {
