+++ src/Contact/Contact.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import './Contact.css';
+
+class Contact extends Component {
+    constructor(props) {
+    //   super(props)
+    //   this.state = {
+    //     active: 'Home'
+    //   }
+    }
+
+    render() {
+        return (
+            <div>This is the Contact component</div>
+        )
+    }
+}
+
+export default Contact;
