+++ src/Home/Home.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import './Home.css';
+
+class Home extends Component {
+    constructor(props) {
+    //   super(props)
+    //   this.state = {
+    //     active: 'Home'
+    //   }
+    }
+
+    render() {
+        return (
+            <div>This is the Home component</div>
+        )
+    }
+}
+
+export default Home;
