+++ src/Settings/Settings.js
@@ -1,0 +1,19 @@
+import React from 'react';
+import './Settings.css';
+
+class Settings extends Component {
+    constructor(props) {
+    //   super(props)
+    //   this.state = {
+    //     active: 'Home'
+    //   }
+    }
+
+    render() {
+        return (
+            <div>This is the Settings component</div>
+        )
+    }
+}
+
+export default Settings;
